+++ src/tools/db/db.php
@@ -4,6 +4,7 @@ namespace Apikor\Tools;
 
 use Vosiz\VaTools\Db\DbConnection      as VaDbConnection;
 use Vosiz\VaTools\Db\DbConnectionConfig as VaDbConnectionConfig;
+use Vosiz\VaTools\Db\DbException;
 
 class DbConnection extends VaDbConnection {
 
@@ -23,6 +24,26 @@ class DbConnection extends VaDbConnection {
     public function LastInsertId() {
 
         return (int)$this->pdo->lastInsertId();
+    }
+
+    /**
+     * Executes raw SQL with optional parameters
+     * @param string $sql
+     * @param array $params
+     * @return bool
+     * @throws DbException
+     */
+    public function Raw(string $sql, array $params = []) {
+
+        try {
+
+            $stmt = $this->pdo->prepare($sql);
+            return $stmt->execute($params);
+
+        } catch(\PDOException $exc) {
+
+            throw new DbException($exc->getMessage());
+        }
     }
 
 }